drop redundant tests
